requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
c3d=0.3.0=pypi_0
ca-certificates=2021.10.26=haa95532_2
certifi=2021.10.8=py38haa95532_0
colorama=0.4.4=pypi_0
configer=1.3.1=pypi_0
configparser=5.2.0=pypi_0
cudatoolkit=11.3.1=h59b6b97_2
freetype=2.10.4=hd328e21_0
freetype-py=2.2.0=pypi_0
h5py=3.6.0=pypi_0
human-body-prior=0.8.5.0=pypi_0
imageio=2.13.2=pypi_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9d=h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
networkx=2.6.3=pypi_0
numpy=1.21.2=py38hfca59bb_0
numpy-base=1.21.2=py38h0829f74_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.4.60=pypi_0
opensim=4.3=pypi_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.3=pypi_0
pillow=8.4.0=py38hd45dc43_0
pip=21.2.2=py38haa95532_0
protobuf=3.19.1=pypi_0
pyglet=1.5.21=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.0.6=pypi_0
pyrender=0.1.45=pypi_0
python=3.8.0=hff0d562_2
pytorch=1.10.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
scikit-image=0.19.0=pypi_0
scipy=1.7.3=pypi_0
setuptools=58.0.4=py38haa95532_0
six=1.16.0=pyhd3eb1b0_0
smplx=0.1.28=pypi_0
sqlite=3.36.0=h2bbff1b_0
tensorboardx=2.4.1=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.10.0=py38_cu113
torchgeometry=0.1.2=pypi_0
torchvision=0.11.1=py38_cu113
tqdm=4.62.3=pypi_0
transforms3d=0.3.1=pypi_0
trimesh=3.9.35=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h62dcd97_0
yacs=0.1.8=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
